人事系統的加班允許小數點

+++ hr/protected/models/AuditWorkForm.php
@@ -404,15 +404,15 @@ class AuditWorkForm extends CFormModel
                 }else{
                     $this->cost_num = 2;
                 }
-                $this->work_cost = ($wage/21.75)*intval($this->log_time)*intval($this->cost_num);
+                $this->work_cost = ($wage/21.75)*floatval($this->log_time)*intval($this->cost_num);
                 break;
             case 1:
-                $this->work_cost = ($wage/(21.75*8))*intval($this->log_time)*2;
+                $this->work_cost = ($wage/(21.75*8))*floatval($this->log_time)*2;
                 $this->start_time .=" ".$this->hours;
                 $this->end_time .=" ".$this->hours_end;
                 break;
             default:
-                $this->work_cost = ($wage/(21.75*8))*intval($this->log_time)*1.5;
+                $this->work_cost = ($wage/(21.75*8))*floatval($this->log_time)*1.5;
                 $this->start_time .=" ".$this->hours;
                 $this->end_time .=" ".$this->hours_end;
         }
